+++ app/config/local/session.php
@@ -18,21 +18,18 @@ return array(
 
     /*
     |--------------------------------------------------------------------------
-    | Password Reminder Language Lines
+    | Default Session Driver
     |--------------------------------------------------------------------------
     |
-    | The following language lines are the default lines which match reasons
-    | that are given by the password broker for a password update attempt
-    | has failed, such as for an invalid token or invalid new password.
+    | This option controls the default session "driver" that will be used on
+    | requests. By default, we will use the lightweight native driver but
+    | you may specify any of the other wonderful drivers provided here.
+    |
+    | Supported: "file", "cookie", "database", "apc",
+    |            "memcached", "redis", "array"
     |
     */
 
-    "password" => "Passwords must be at least six characters and match the confirmation.",
-
-    "user"     => "We can't find a user with that e-mail address.",
-
-    "token"    => "This password reset token is invalid.",
-
-    "sent"     => "Password reminder sent!",
+    'driver' => 'file'
 
 );