+++ src/Navigation/Factory.php
@@ -1,0 +1,139 @@
+<?php
+
+/**
+ * This file is part of Bootstrap CMS by Graham Campbell.
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU Affero General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU Affero General Public License for more details.
+ */
+
+namespace GrahamCampbell\BootstrapCMS\Navigation;
+
+use GrahamCampbell\Navigation\Navigation;
+use GrahamCampbell\Credentials\Credentials;
+
+/**
+ * This is the navigation factory class.
+ *
+ * @package    Bootstrap-CMS
+ * @author     Graham Campbell
+ * @copyright  Copyright (C) 2013-2014  Graham Campbell
+ * @license    https://github.com/GrahamCampbell/Bootstrap-CMS/blob/master/LICENSE.md
+ * @link       https://github.com/GrahamCampbell/Bootstrap-CMS
+ */
+class Factory
+{
+    /**
+     * The credentials instance.
+     *
+     * @var \GrahamCampbell\Credentials\Credentials
+     */
+    protected $credentials;
+
+    /**
+     * The navigation instance.
+     *
+     * @var \GrahamCampbell\Navigation\Navigation
+     */
+    protected $navigation;
+
+    /**
+     * The platform name.
+     *
+     * @var string
+     */
+    protected $name;
+
+    /**
+     * The inverse navigation.
+     *
+     * @var bool
+     */
+    protected $inverse;
+
+   /**
+     * Create a new instance.
+     *
+     * @param  \GrahamCampbell\Credentials\Credentials  $credentials
+     * @param  \GrahamCampbell\Navigation\Navigation  $navigation
+     * @param  string  $name
+     * @param  bool  $inverse
+     * @return void
+     */
+    public function __construct(Credentials $credentials, Navigation $navigation, $name, $inverse)
+    {
+        $this->credentials = $credentials;
+        $this->navigation = $navigation;
+        $this->name = $name;
+        $this->inverse = $inverse;
+    }
+
+    /**
+     * Create a navigation bar.
+     *
+     * @param  string  $type
+     * @return void
+     */
+    public function make($type = 'default')
+    {
+        if ($this->credentials->check()) {
+            if ($type === 'admin') {
+                if ($this->credentials->hasAccess('admin')) {
+                    // the requested type is admin, and the user is an admin
+                    return $this->navigation->getHTML('admin', 'admin', array(
+                        'title' => 'Admin Panel',
+                        'side' => $this->credentials->getUser()->email,
+                        'inverse' => $this->inverse
+                    ));
+                } else {
+                    // the requested type is admin, and the user is NOT an admin
+                    return $this->navigation->getHTML('default', 'default', array(
+                        'title' => $this->name,
+                        'side' => $this->credentials->getUser()->email,
+                        'inverse' => $this->inverse
+                    ));
+                }
+            } else {
+                // the requested type is default, and the user is logged in
+                return $this->navigation->getHTML('default','default', array(
+                    'title' => $this->name,
+                    'side' => $this->credentials->getUser()->email,
+                    'inverse' => $this->inverse
+                ));
+            }
+        } else {
+            // the requested type is default, and the user is NOT logged in
+            return $this->navigation->getHTML('default', false, array(
+                'title' => $this->name,
+                'inverse' => $this->inverse
+            ));
+        }
+    }
+
+    /**
+     * Return the credentials instance.
+     *
+     * @return \GrahamCampbell\Credentials\Credentials
+     */
+    public function getCredentials()
+    {
+        return $this->credentials;
+    }
+
+    /**
+     * Return the navigation instance.
+     *
+     * @return \GrahamCampbell\Navigation\Navigation
+     */
+    public function getNavigation()
+    {
+        return $this->navigation;
+    }
+}